+++ piglegsurgeryweb/uploader/urls.py
@@ -59,4 +59,6 @@ urlpatterns = [
     path("delete_media_file/<int:filename_id>/", views.delete_media_file, name="delete_media_file"),
     path("download_sample_image", views.download_sample_image, name="download_sample_image"),
     path("common_review", views.common_review, name="common_review"),
+    path("collections/", views.collections, name="collections"),
+    path("run_collection/<int:collection_id>/", views.run_collection, name="run_collection"),
 ]